+++ src/App.css
@@ -1,4 +1,4 @@
 ul li {
-  list-style-type: armenian;
+  list-style-type: upper-latin;
   color: red;
 }
