+++ src/App.tsx
@@ -1,34 +1,14 @@
-import React from 'react';
+// import React from 'react';
 import './App.css';
+import Header from './Header';
+import Technologies from './Technologies';
+
 
 const App = () => {
   return (
     <div>
       <Header />
       <Technologies />
-    </div>
-  );
-}
-
-const Header = () => {
-  return (
-    <div>
-      <a href='#'>Home</a>
-      <a href='#'>News Feed</a>
-      <a href='#'>Message</a>
-    </div>
-  );
-}
-
-const Technologies = () => {
-  return (
-    <div>
-      <ul>
-        <li>js</li>
-        <li>css</li>
-        <li>react</li>
-        <li>html</li>
-      </ul>
     </div>
   );
 }
