+++ src/Technologies.js
@@ -1,0 +1,14 @@
+const Technologies = () => {
+    return (
+        <div>
+            <ul>
+                <li>js</li>
+                <li>css</li>
+                <li>react</li>
+                <li>html</li>
+            </ul>
+        </div>
+    );
+}
+
+export default Technologies;
